refactor: ExchangeApiClient interface
