+++ recording-oracle/src/modules/exchange/index.ts
@@ -1,11 +1,11 @@
 export { ExchangeModule } from './exchange.module';
 export { ExchangeApiClientFactory } from './exchange-api-client-factory';
 export { ExchangeApiClientError } from './errors';
+export type { Order, Trade } from './types';
+
 export type {
   ExchangeApiClient,
   ExchangeApiClientInitOptions,
-  Order,
-  Trade,
-} from './types';
+} from './exchange-api-client.interface';
 
 export { TakerOrMakerFlag, TradingSide } from './types';
