+++ recording-oracle/src/modules/exchange/types.ts
@@ -22,6 +22,7 @@ export type Order = Pick<
   | 'filled'
   | 'cost'
 >;
+
 export type Trade = Pick<
   CcxtTrade,
   | 'id'
@@ -33,20 +34,3 @@ export type Trade = Pick<
   | 'amount'
   | 'cost'
 >;
-
-export type ExchangeApiClientInitOptions = {
-  apiKey: string;
-  secret: string;
-};
-
-export interface ExchangeApiClient {
-  readonly exchangeName: string;
-
-  checkRequiredCredentials(): boolean;
-
-  checkRequiredAccess(): Promise<boolean>;
-
-  fetchOpenOrders(symbol: string, since: number): Promise<Order[]>;
-
-  fetchMyTrades(symbol: string, since: number): Promise<Trade[]>;
-}
